Reorganize checksum specs into subdirectory   Split up by class under testing Add specs for md5sum value validator. Add integration test for calculating checksums.   Add mismatched_checksum.zip fixture to serve as failing example. Implement Chesum Expected Value class   Add class to module includes
